requirements.txt
fastapi
uvicorn[standard]
jinja2
pyyaml
openai
python-dotenv
